requirements.txt
requests==2.30.0

streamlit==1.23.1
Pillow==9.5.0
jdatetime==4.1.1
pandas==2.0.1